fix(operations): resolve compilation errors blocking DLQ retry functionality
Fixed multiple compilation errors that prevented the newly integrated DLQ retry endpoint from functioning. Added missing error variants, corrected circuit breaker method calls, resolved type mismatches in database pool stats, and fixed borrowing conflicts in price cache. The DLQ retry route is now fully operational and the operator compiles cleanly.

Technical fixes:
- Added BadRequest and ServiceUnavailable error variants
- Fixed circuit breaker state comparison using current_state()
- Resolved u32/usize type casting in database pool statistics
- Fixed price cache borrowing conflict with early clone
- Moved misplaced get_db_query_stats() method to correct impl block
- Updated deprecated Prometheus API calls
- Added missing price_cache field to ApiState initialization

+++ operator/src/db_abstraction/sqlite.rs
@@ -95,8 +95,8 @@ impl Database for SqliteBackend {
     async fn get_pool_stats(&self) -> AppResult<super::PoolStats> {
         use super::PoolStats;
 
-        let max_connections = self.pool.size();
-        let idle_connections = self.pool.num_idle();
+        let max_connections = self.pool.size() as usize;
+        let idle_connections = self.pool.num_idle() as usize;
         let active_connections = max_connections.saturating_sub(idle_connections);
 
         let utilization_percent = if max_connections > 0 {
